+++ Source/TrinityFlow/Private/Enemy/ShieldedTankEnemy.cpp
@@ -5,20 +5,10 @@
 AShieldedTankEnemy::AShieldedTankEnemy()
 {
     bIsAreaDamage = false;
+    EnemyStatsID = "ShieldedTankEnemy";
 }
 
 void AShieldedTankEnemy::SetupEnemy()
 {
-    // Tags: havesoul, armored, shielded
-    if (TagComponent)
-    {
-        TagComponent->SetTags(ECharacterTag::HaveSoul | ECharacterTag::Armored | ECharacterTag::Shielded);
-    }
-
-    // Resources: health: 300, attackpoint: 20, defencepoint: 20
-    if (HealthComponent)
-    {
-        FCharacterResources Resources(300.0f, 20.0f, 20.0f);
-        HealthComponent->SetResources(Resources);
-    }
+    Super::SetupEnemy();
 }