+++ Source/TrinityFlow/Private/Enemy/ShieldedTankRobotEnemy.cpp
@@ -5,20 +5,10 @@
 AShieldedTankRobotEnemy::AShieldedTankRobotEnemy()
 {
     bIsAreaDamage = false;
+    EnemyStatsID = "ShieldedTankRobotEnemy";
 }
 
 void AShieldedTankRobotEnemy::SetupEnemy()
 {
-    // Tags: mechanical, armored, shielded
-    if (TagComponent)
-    {
-        TagComponent->SetTags(ECharacterTag::Mechanical | ECharacterTag::Armored | ECharacterTag::Shielded);
-    }
-
-    // Resources: health: 300, attackpoint: 30, defencepoint: 40
-    if (HealthComponent)
-    {
-        FCharacterResources Resources(300.0f, 40.0f, 30.0f);
-        HealthComponent->SetResources(Resources);
-    }
+    Super::SetupEnemy();
 }