+++ Source/TrinityFlow/Private/Enemy/StandardEnemy.cpp
@@ -4,21 +4,11 @@
 
 AStandardEnemy::AStandardEnemy()
 {
-    bIsAreaDamage = false;
+    EnemyStatsID = "StandardEnemy";
 }
 
 void AStandardEnemy::SetupEnemy()
 {
-    // Tags: havesoul
-    if (TagComponent)
-    {
-        TagComponent->SetTags(ECharacterTag::HaveSoul);
-    }
-
-    // Resources: health: 100, attackpoint: 10, defencepoint: 0
-    if (HealthComponent)
-    {
-        FCharacterResources Resources(100.0f, 0.0f, 10.0f);
-        HealthComponent->SetResources(Resources);
-    }
+    // Call parent to load stats from data
+    Super::SetupEnemy();
 }